Revert "Refactor: Add no reminder label"
