+++ quick-reminder/Model/ReminderList/ReminderList.swift
@@ -19,8 +19,6 @@ extension Notification.Name {
 /// Reminder配列の管理を行うためのメソッド。
 protocol ReminderListProtocol {
     var notificationCenter: NotificationCenter { get }
-    /// 格納しているReminderの配列。
-    var reminders: [Reminder] { get }
     /// Reminderリストの要素数。
     var count: Int { get }
     /// Reminderリストが空かを表すブール値。
@@ -28,6 +26,9 @@ protocol ReminderListProtocol {
     
     /// リポジトリからデータをフェッチすることでリストを初期化する。
     func fetchReminders()
+    
+    /// 与えられたインデックスのReminderを返す。
+    func getReminder(index: Int) -> Reminder
     
     /// 与えられたReminderがどのindexで管理されているかを返す。
     ///
@@ -69,7 +70,7 @@ final class ReminderList: ReminderListProtocol {
     private let sorter: ReminderSorterProtocol!
     private let validator: ReminderListValidatorProtocol!
     
-    private(set) var reminders: [Reminder] = [] {
+    private var reminders: [Reminder] = [] {
         didSet {
             reminders = sorter.sorted(reminders: reminders)
         }
@@ -88,6 +89,10 @@ final class ReminderList: ReminderListProtocol {
     
     func fetchReminders() {
         reminders = repository.getAllReminders()
+    }
+    
+    func getReminder(index: Int) -> Reminder {
+        reminders[index]
     }
     
     func getIndex(reminder: Reminder) throws -> Int {
