+++ quick-reminder/SceneDelegate.swift
@@ -20,11 +20,6 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         
         let rootVC = ReminderListViewController(
             dependency: .init(
-                dateProvider: DateProvider()
-            )
-        )
-        let vcPresenter = ReminderListPresenter(
-            dependency: .init(
                 reminderList: ReminderList(
                     repository: useUITestMockRepository ? UITestMockReminderRepository() : ReminderRepository(),
                     sorter: ReminderSorter(),
@@ -34,12 +29,10 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
                 notificationDateCalculator: NotificationDateCalculator(dateProvider: DateProvider()),
                 dateProvider: DateProvider(),
                 oldReminderRemover: OldReminderRemover(dateProvider: DateProvider())
-            ),
-            view: rootVC
+            )
         )
-        rootVC.inject(presenter: vcPresenter)
-        
         window?.rootViewController = UINavigationController(rootViewController: rootVC)
+        
         window?.makeKeyAndVisible()
     }
